Refactor all location tests into 1
